idek anymore

+++ src/battle/decide_move.cpp
@@ -4,5 +4,5 @@
 // place to implement decision making
 void Battle::decide_move(Team &team){
     team.prev_movechoice = team.movechoice;
-    team.movechoice = team.move_options[this->transition.random(static_cast<int>(team.move_options.size()))];
+    team.movechoice = team.move_options[this->transition.random(team.move_options.size())];
 }
